+++ dev/Basic/geospatial/Link.hpp
@@ -42,7 +42,7 @@ public:
 
 	///Return the length of this Link, which is the sum of all RoadSegments
 	/// in the forward (if isForward is true) direction.
-	double getLength(bool isForward) const;
+	int getLength(bool isForward) const;
 
 	///Return the RoadSegments which make up this Link, in either the forward
 	/// (if isForward is true) or reverse direction.
